Extract useCollabCursors hook for the shared collab-cursor wiring
The text block and footnote editors each repeated the same three-part
wiring: pull awareness out of useYJSValue, add remoteCursorPlugin to the
plugin list, and render a RemoteCursors overlay. Consolidate that into a
single useCollabCursors(entityID) hook (which also now owns useYJSValue and
memoizes the plugin), returning { yText, cursorPlugin, overlay } so a caller
just drops cursorPlugin into its plugins and renders overlay. The plugin and
overlay can no longer drift apart between the two call sites.
