+++ components/Blocks/TextBlock/mountProsemirror.ts
@@ -1,12 +1,9 @@
-import { useLayoutEffect, useRef, useEffect, useState } from "react";
+import { useLayoutEffect, useRef } from "react";
 import { EditorState, Transaction } from "prosemirror-state";
 import { EditorView } from "prosemirror-view";
 import { baseKeymap } from "prosemirror-commands";
 import { keymap } from "prosemirror-keymap";
 import { ySyncPlugin } from "y-prosemirror";
-import * as Y from "yjs";
-import { Awareness } from "y-protocols/awareness";
-import * as base64 from "base64-js";
 import { Replicache } from "replicache";
 import { produce } from "immer";
 
@@ -30,9 +27,7 @@ import { useFootnotePopoverStore } from "components/Footnotes/FootnotePopover";
 import {
   useLinkPopoverStore,
 } from "components/LinkPopover";
-import { useYjsRealtime, YjsRealtimeConnection } from "src/yjsRealtime";
-import { useIdentityData } from "components/IdentityProvider";
-import { remoteCursorPlugin } from "./remoteCursorPlugin";
+import { useCollabCursors } from "./useCollabCursors";
 
 export function useMountProsemirror({
   props,
@@ -45,7 +40,7 @@ export function useMountProsemirror({
   let rep = useReplicache();
   let mountRef = useRef<HTMLPreElement | null>(null);
   const repRef = useRef<Replicache<ReplicacheMutators> | null>(null);
-  let { yText: value, awareness } = useYJSValue(entityID);
+  let { yText: value, cursorPlugin, overlay } = useCollabCursors(entityID);
   let entity_set = useEntitySetContext();
   let alignment =
     useEntity(entityID, "block/text-alignment")?.data.value || "left";
@@ -70,7 +65,7 @@ export function useMountProsemirror({
       schema: schema,
       plugins: [
         ySyncPlugin(value),
-        remoteCursorPlugin(awareness),
+        cursorPlugin,
         keymap(km),
         inputrules(propsRef, repRef, openMentionAutocomplete),
         keymap(baseKeymap),
@@ -280,8 +275,8 @@ export function useMountProsemirror({
         };
       });
     }
-  }, [entityID, parent, value, awareness, handlePaste, rep]);
-  return { mountRef, actionTimeout, awareness };
+  }, [entityID, parent, value, cursorPlugin, handlePaste, rep]);
+  return { mountRef, actionTimeout, overlay };
 }
 
 export function trackUndoRedo(
@@ -308,81 +303,4 @@ export function trackUndoRedo(
 
     undoManager.add({ undo, redo });
   }
-}
-
-export function useYJSValue(entityID: string) {
-  const [ydoc] = useState(() => new Y.Doc());
-  const docStateFromReplicache = useEntity(entityID, "block/text");
-  let rep = useReplicache();
-  let realtime = useYjsRealtime();
-  let { identity } = useIdentityData();
-  const [yText] = useState(() => ydoc.getXmlFragment("prosemirror"));
-  const [awareness] = useState(() => new Awareness(ydoc));
-
-  if (docStateFromReplicache) {
-    const update = base64.toByteArray(docStateFromReplicache.data.value);
-    Y.applyUpdate(ydoc, update);
-  }
-
-  let userName =
-    identity?.bsky_profiles?.record?.displayName ||
-    identity?.bsky_profiles?.handle ||
-    "Anonymous";
-  useEffect(() => {
-    // a stable hue offset per client; each peer derives the actual cursor
-    // color from their own theme's accent (see collabCursor.ts)
-    awareness.setLocalStateField("user", {
-      name: userName,
-      hue: (ydoc.clientID % 8) * 45,
-    });
-  }, [awareness, userName, ydoc]);
-
-  // Broadcast local edits and cursor positions to connected peers, and apply
-  // theirs as they arrive. Cursors only ever travel over the channel; doc
-  // updates also flow through replicache below.
-  useEffect(() => {
-    if (!realtime) return;
-    return realtime.register(entityID, ydoc, awareness);
-  }, [realtime, entityID, ydoc, awareness]);
-
-  useEffect(() => {
-    return () => {
-      awareness.destroy();
-    };
-  }, [awareness]);
-
-  useEffect(() => {
-    if (!rep.rep) return;
-    let timeout = null as null | number;
-    const updateReplicache = async () => {
-      const update = Y.encodeStateAsUpdate(ydoc);
-      await rep.rep?.mutate.assertFact({
-        //These undos are handled above in the Prosemirror context
-        ignoreUndo: true,
-        entity: entityID,
-        attribute: "block/text",
-        data: {
-          value: base64.fromByteArray(update),
-          type: "text",
-        },
-      });
-    };
-    const f = async (events: Y.YEvent<any>[], transaction: Y.Transaction) => {
-      // Transactions with no origin come from replicache itself, and ones
-      // originating from the realtime channel are persisted by the peer that
-      // authored them — only local edits should be written back.
-      if (!transaction.origin) return;
-      if (transaction.origin instanceof YjsRealtimeConnection) return;
-      if (timeout) clearTimeout(timeout);
-      timeout = window.setTimeout(async () => {
-        updateReplicache();
-      }, 300);
-    };
-
-    yText.observeDeep(f);
-    return () => {
-      yText.unobserveDeep(f);
-    };
-  }, [yText, entityID, rep, ydoc]);
-  return { yText, awareness };
 }
